Createa new packageList component
