+++ src/pages/ExtensionsPage.vue
@@ -57,246 +57,17 @@ limitations under the License.
       <div v-if="customPackages.length === 0" class="text-grey-7 q-mt-sm">
         No custom extensions added yet.
       </div>
-      <q-list class="cluster-list">
-        <div
-          v-for="(sessionPackage, index) in customPackages"
-          :key="'custom-' + index"
-        >
-          <q-item dense class="q-px-none">
-            <q-item-section>
-              <q-expansion-item>
-                <template #header>
-                  <q-item-section avatar class="q-pr-none">
-                    <q-icon
-                      :class="{
-                        'cursor-pointer':
-                          iconName(sessionPackage.pkg.id) == 'error' ||
-                          iconName(sessionPackage.pkg.id) == 'warning'
-                      }"
-                      :name="iconName(sessionPackage.pkg.id)"
-                      :color="iconColor(sessionPackage.pkg.id)"
-                      size="1.5em"
-                      @click="() => handleIconClick(sessionPackage.pkg.id)"
-                    />
-                  </q-item-section>
-                  <div class="q-my-auto q-item__label q-item__label__popup">
-                    <strong>{{ getFileName(sessionPackage.pkg.path) }}</strong>
-                  </div>
-                  <q-space />
-                  <q-btn
-                    class="q-mx-xl"
-                    label="Delete"
-                    icon="delete"
-                    flat
-                    dense
-                    @click.stop="deletePackage(sessionPackage)"
-                    :disable="sessionPackage.sessionPackage.required"
-                  />
-                </template>
-                <q-card>
-                  <q-card-section>
-                    <div class="q-mx-lg q-px-lg">
-                      <strong> Full File path:</strong>
-                      {{ sessionPackage.pkg.path }} <br />
-                      <strong> Package Type:</strong>
-                      {{ sessionPackage.pkg.type }} <br />
-                      <strong> Version: </strong
-                      >{{ sessionPackage.pkg.version }} <br />
-                      <strong> Required:</strong>
-                      {{
-                        sessionPackage.sessionPackage.required
-                          ? 'True'
-                          : 'False'
-                      }}
-                    </div>
-                  </q-card-section>
-                </q-card>
-              </q-expansion-item>
-            </q-item-section>
-            <q-dialog v-model="dialogData[sessionPackage.pkg.id]">
-              <q-card>
-                <q-card-section>
-                  <div class="row items-center">
-                    <div class="col-1">
-                      <q-icon
-                        :name="iconName(sessionPackage.pkg.id)"
-                        :color="iconColor(sessionPackage.pkg.id)"
-                        size="2em"
-                      ></q-icon>
-                    </div>
-                    <div class="text-h6 col">
-                      {{ sessionPackage.pkg.path }}
-                    </div>
-                    <div class="col-1 text-right">
-                      <q-btn dense flat icon="close" v-close-popup>
-                        <q-tooltip>Close</q-tooltip>
-                      </q-btn>
-                    </div>
-                  </div>
-                  <div v-if="notisData[sessionPackage.pkg.id]?.hasError">
-                    <div
-                      class="text-h6"
-                      style="margin-top: 15px; padding-left: 20px"
-                    >
-                      Errors
-                    </div>
-                    <ul>
-                      <li
-                        v-for="(error, index) in populateNotifications(
-                          sessionPackage.pkg.id,
-                          'ERROR'
-                        )"
-                        :key="'error' + index"
-                        style="margin-bottom: 10px"
-                      >
-                        {{ error.message }}
-                      </li>
-                    </ul>
-                  </div>
-                  <div v-if="notisData[sessionPackage.pkg.id]?.hasWarning">
-                    <div
-                      class="text-h6"
-                      style="margin-top: 15px; padding-left: 20px"
-                    >
-                      Warnings
-                    </div>
-                    <ul>
-                      <li
-                        v-for="(warning, index) in populateNotifications(
-                          sessionPackage.pkg.id,
-                          'WARNING'
-                        )"
-                        :key="index"
-                        style="margin-bottom: 10px"
-                      >
-                        {{ warning.message }}
-                      </li>
-                    </ul>
-                  </div>
-                </q-card-section>
-              </q-card>
-            </q-dialog>
-          </q-item>
-        </div>
-      </q-list>
+      <PackagesList :packages="customPackages" :notisData="notisData" v-else />
     </q-card-section>
     <q-card-section class="q-pt-none">
       <div class="row items-center">
         <strong>Built-in ZCL packages</strong>
       </div>
-      <q-list class="cluster-list">
-        <div
-          v-for="(sessionPackage, index) in builtInPackages"
-          :key="'builtin-' + index"
-        >
-          <q-item dense class="q-px-none">
-            <q-item-section>
-              <q-expansion-item>
-                <template #header>
-                  <q-item-section avatar class="q-pr-none">
-                    <q-icon
-                      :class="{
-                        'cursor-pointer':
-                          iconName(sessionPackage.pkg.id) == 'error' ||
-                          iconName(sessionPackage.pkg.id) == 'warning'
-                      }"
-                      :name="iconName(sessionPackage.pkg.id)"
-                      :color="iconColor(sessionPackage.pkg.id)"
-                      size="1.5em"
-                      @click="() => handleIconClick(sessionPackage.pkg.id)"
-                    />
-                  </q-item-section>
-                  <div class="q-my-auto q-item__label q-item__label__popup">
-                    <strong>{{ getFileName(sessionPackage.pkg.path) }}</strong>
-                  </div>
-                  <q-space />
-                </template>
-                <q-card>
-                  <q-card-section>
-                    <div class="q-mx-lg q-px-lg">
-                      <strong> Full File path:</strong>
-                      {{ sessionPackage.pkg.path }} <br />
-                      <strong> Package Type:</strong>
-                      {{ sessionPackage.pkg.type }} <br />
-                      <strong> Version: </strong
-                      >{{ sessionPackage.pkg.version }} <br />
-                      <strong> Required:</strong>
-                      {{
-                        sessionPackage.sessionPackage.required
-                          ? 'True'
-                          : 'False'
-                      }}
-                    </div>
-                  </q-card-section>
-                </q-card>
-              </q-expansion-item>
-            </q-item-section>
-            <q-dialog v-model="dialogData[sessionPackage.pkg.id]">
-              <q-card>
-                <q-card-section>
-                  <div class="row items-center">
-                    <div class="col-1">
-                      <q-icon
-                        :name="iconName(sessionPackage.pkg.id)"
-                        :color="iconColor(sessionPackage.pkg.id)"
-                        size="2em"
-                      ></q-icon>
-                    </div>
-                    <div class="text-h6 col">
-                      {{ sessionPackage.pkg.path }}
-                    </div>
-                    <div class="col-1 text-right">
-                      <q-btn dense flat icon="close" v-close-popup>
-                        <q-tooltip>Close</q-tooltip>
-                      </q-btn>
-                    </div>
-                  </div>
-                  <div v-if="notisData[sessionPackage.pkg.id]?.hasError">
-                    <div
-                      class="text-h6"
-                      style="margin-top: 15px; padding-left: 20px"
-                    >
-                      Errors
-                    </div>
-                    <ul>
-                      <li
-                        v-for="(error, index) in populateNotifications(
-                          sessionPackage.pkg.id,
-                          'ERROR'
-                        )"
-                        :key="'error' + index"
-                        style="margin-bottom: 10px"
-                      >
-                        {{ error.message }}
-                      </li>
-                    </ul>
-                  </div>
-                  <div v-if="notisData[sessionPackage.pkg.id]?.hasWarning">
-                    <div
-                      class="text-h6"
-                      style="margin-top: 15px; padding-left: 20px"
-                    >
-                      Warnings
-                    </div>
-                    <ul>
-                      <li
-                        v-for="(warning, index) in populateNotifications(
-                          sessionPackage.pkg.id,
-                          'WARNING'
-                        )"
-                        :key="index"
-                        style="margin-bottom: 10px"
-                      >
-                        {{ warning.message }}
-                      </li>
-                    </ul>
-                  </div>
-                </q-card-section>
-              </q-card>
-            </q-dialog>
-          </q-item>
-        </div>
-      </q-list>
+      <PackagesList
+        :packages="builtInPackages"
+        :notisData="notisData"
+        builtIn
+      />
     </q-card-section>
   </PreferencePageLayout>
 </template>
@@ -306,13 +77,15 @@ import CommonMixin from '../util/common-mixin'
 import rendApi from '../../src-shared/rend-api.js'
 import restApi from '../../src-shared/rest-api.js'
 import PreferencePageLayout from '../layouts/PreferencePageLayout.vue'
+import PackagesList from '../components/PackagesList.vue'
 const observable = require('../util/observable.js')
 import { Notify } from 'quasar'
 
 export default {
   mixins: [CommonMixin],
   components: {
-    PreferencePageLayout
+    PreferencePageLayout,
+    PackagesList
   },
   watch: {
     packages(newPackages) {
@@ -337,10 +110,6 @@ export default {
     }
   },
   methods: {
-    getFileName(path) {
-      let fileName = path.match(/[^/]+$/)
-      return fileName.length > 0 ? fileName[0] : path
-    },
     browseForFile() {
       window[rendApi.GLOBAL_SYMBOL_NOTIFY](rendApi.notifyKey.fileBrowse, {
         context: 'customXml',
@@ -377,18 +146,9 @@ export default {
           let packageId = packageFile.pkg.id
           if (packageId) {
             this.getPackageNotifications(packageId)
-            this.dialogData[packageId] = false
           }
         })
       }
-    },
-    async deletePackage(packageToDelete) {
-      await this.$store.dispatch(
-        'zap/deleteSessionPackage',
-        packageToDelete.sessionPackage
-      )
-      await this.$store.dispatch('zap/updateClusters')
-      await this.$store.dispatch('zap/updateAtomics')
     },
     async getPackageNotifications(packageId) {
       this.$serverGet(
@@ -409,36 +169,8 @@ export default {
             currentPackage.warnings.push(notification)
           }
         })
-        this.notisData[packageId] = currentPackage
+        this.$set(this.notisData, packageId, currentPackage)
       })
-    },
-    iconName(packageId) {
-      if (this.notisData[packageId]?.hasError) {
-        return 'error'
-      } else if (this.notisData[packageId]?.hasWarning) {
-        return 'warning'
-      } else {
-        return 'check_circle'
-      }
-    },
-    iconColor(packageId) {
-      if (this.notisData[packageId]?.hasError) {
-        return 'red'
-      } else if (this.notisData[packageId]?.hasWarning) {
-        return 'orange'
-      } else {
-        return 'green'
-      }
-    },
-    handleIconClick(packageId) {
-      let iconName = this.iconName(packageId)
-      if (iconName === 'error' || iconName === 'warning') {
-        this.dialogData[packageId] = true
-      }
-    },
-    populateNotifications(packageId, type) {
-      let key = type == 'ERROR' ? 'errors' : 'warnings'
-      return this.notisData[packageId][key]
     },
     // Custom xml currently not supported for multi-protocol
     enableExtensionsWarning() {
@@ -467,8 +199,7 @@ export default {
     return {
       packageToLoad: '',
       error: null,
-      notisData: {},
-      dialogData: {}
+      notisData: {}
     }
   }
 }
